Don't force the user out of touch mode.
Commit b9f48b2 made the SpeedDialView call requestFocusFromTouch() for
the first FAB upon toggle(). However, requestFocusFromTouch() will
force the user out of touch mode and the direct impact on UI is that
the first FAB gains a focused effect and stands out from other FABs.

This isn't the desired behavior because buttons, unlike text fields,
should never gain focus in touch mode, and this is unexpected for most
users (without a keyboard/trackball), both the behavior and the
look-and-feel. The official guidance on touch mode can be found at
https://android-developers.googleblog.com/2008/12/touch-mode.html ,
and the tl;dr is that "Do not try to keep the focus or selection in
touch mode", plus the implied don't force users out of touch mode as
users only enter touch mode when they start to use a keyboard or
trackball. If the user is already using a keyboard or trackball, then
they won't be in touch mode and the regular requestFocus() will just
work for them. This is also consistent with the speed dial
implementation that can be found in Google apps and I assume that they
have proper a11y support.